Поправил ошибки линтера
