+++ src/components/sort-options/sort-option.ts
@@ -1,0 +1,6 @@
+export enum SortOption {
+  Popular = 'Popular',
+  PriceLowToHigh = 'Price: low to high',
+  PriceHighToLow = 'Price: high to low',
+  TopRatedFirst = 'Top rated first'
+}
